requirements.txt
absl-py==0.7.1
astor==0.8.0
astroid==2.2.5
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
entrypoints==0.3
fire==0.1.3
flake8==3.7.7
Flask==1.0.3
Flask-Cors==3.0.8
gast==0.2.2
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-storage==1.16.1
google-pasta==0.1.7
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.21.1
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.4
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
python-dotenv==0.10.3
pytz==2019.1
regex==2017.4.5
requests==2.21.0
rsa==4.0
six==1.12.0
# tensorboard==1.14.0
# tensorflow==1.14.0
# tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.31.1
typed-ast==1.4.0
urllib3==1.24.3
uWSGI==2.0.18
Werkzeug==0.15.4
wrapt==1.11.2

